userRating Float형으로 변경

+++ 2005/chuuuul/ArchitectureStudy/app/src/main/java/com/example/architecture/activity/Activity/MovieVO.kt
@@ -5,5 +5,5 @@ data class MovieVO(
         val title : String,
         val pubDate: String,
         val link : String,
-        val userRating : Int
+        val userRating: Float
 )
